Update "number of rows" when setting dvData (#1)
When setting the `dvData` property of the DataView the `RenderView` handler will be called. This handler relies on the `viewProp["number of rows"]` property which must accurately reflect the number of rows. By intercepting the `setProp` and adjusting the number of rows to the number of elements in the array `RenderView` works correctly.